Cleanup and refactoring of db/core code
