+++ production/db/core/inc/db_internal_types.hpp
@@ -127,10 +127,9 @@ struct txn_log_t
     friend std::ostream& operator<<(std::ostream& os, const txn_log_t& l)
     {
         os << "record_count: " << l.record_count << std::endl;
-        const log_record_t* const lr_start = static_cast<const log_record_t*>(l.log_records);
-        for (const log_record_t* lr = lr_start; lr < lr_start + l.record_count; ++lr)
+        for (size_t i = 0; i < l.record_count; ++i)
         {
-            os << *lr;
+            os << l.log_records[i];
         }
         os << std::endl;
         return os;
@@ -185,295 +184,5 @@ struct id_index_t
     size_t hash_node_count;
     hash_node_t hash_nodes[c_hash_buckets + c_max_locators];
 };
-
-// Base class abstracting common functionality for mapped_data_t and mapped_log_t classes.
-template <typename T>
-class base_mapped_data_t
-{
-public:
-    base_mapped_data_t()
-    {
-        clear();
-    }
-
-    // Copy semantics is disabled and moves should be performed via reset().
-    base_mapped_data_t(const base_mapped_data_t& other) = delete;
-    base_mapped_data_t(base_mapped_data_t&& other) = delete;
-    base_mapped_data_t& operator=(const base_mapped_data_t& rhs) = delete;
-    base_mapped_data_t& operator=(base_mapped_data_t&& rhs) = delete;
-
-    ~base_mapped_data_t()
-    {
-        close();
-    }
-
-    // Stops tracking any data and reverts back to uninitialized state.
-    void clear()
-    {
-        m_is_set = false;
-        m_fd = -1;
-        m_data = nullptr;
-        m_mapped_data_size = 0;
-    }
-
-    // Transfers data tracked by another instance into this instance.
-    void reset(base_mapped_data_t<T>& other)
-    {
-        common::retail_assert(
-            !m_is_set,
-            "A set base_mapped_data_t instance should not take ownership of another instance!");
-        common::retail_assert(
-            other.m_is_set,
-            "An unset base_mapped_data_t instance should not take ownership of another unset instance!");
-
-        m_is_set = other.m_is_set;
-        m_fd = other.m_fd;
-        m_data = other.m_data;
-        m_mapped_data_size = other.m_mapped_data_size;
-
-        other.clear();
-    }
-
-    // Unmaps the data and closes the file descriptor, if one is tracked.
-    // Reverts back to uninitialized state.
-    // This permits manual cleanup, before instance destruction time.
-    // Can be called repeatedly.
-    void close()
-    {
-        common::unmap_fd_data(m_data, m_mapped_data_size);
-        m_mapped_data_size = 0;
-
-        common::close_fd(m_fd);
-
-        m_is_set = false;
-    }
-
-    T* data()
-    {
-        return m_data;
-    }
-
-    int fd()
-    {
-        return m_fd;
-    }
-
-    bool is_set()
-    {
-        return m_is_set;
-    }
-
-protected:
-    bool m_is_set;
-    int m_fd;
-    T* m_data;
-
-    // This is used to track the mapped data size, so we can call unmap_fd_data()/munmap() with the same value.
-    size_t m_mapped_data_size;
-};
-
-// This class abstracts the server and client operations with memory-mapped data.
-// T indicates the type of data structure that is managed by an instance of this class.
-template <typename T>
-class mapped_data_t : public base_mapped_data_t<T>
-{
-public:
-    mapped_data_t() = default;
-
-    // Copy semantics is disabled and moves should be performed via reset().
-    mapped_data_t(const mapped_data_t& other) = delete;
-    mapped_data_t(mapped_data_t&& other) = delete;
-    mapped_data_t& operator=(const mapped_data_t& rhs) = delete;
-    mapped_data_t& operator=(mapped_data_t&& rhs) = delete;
-
-    ~mapped_data_t() = default;
-
-    // Creates a memory-mapping for a data structure.
-    void create(const char* name)
-    {
-        common::retail_assert(
-            !this->m_is_set,
-            "Calling create() on an already set mapped_data_t instance!");
-
-        this->m_fd = ::memfd_create(name, MFD_ALLOW_SEALING);
-        if (this->m_fd == -1)
-        {
-            common::throw_system_error("memfd_create() failed in mapped_data_t::create()!");
-        }
-
-        this->m_mapped_data_size = sizeof(T);
-
-        common::truncate_fd(this->m_fd, this->m_mapped_data_size);
-
-        // Note that unless we supply the MAP_POPULATE flag to mmap(), only the
-        // pages we actually use will ever be allocated. However, Linux may refuse
-        // to allocate the virtual memory we requested if overcommit is disabled
-        // (i.e., /proc/sys/vm/overcommit_memory = 2). Using MAP_NORESERVE (don't
-        // reserve swap space) should ensure that overcommit always succeeds, unless
-        // it is disabled. We may need to document the requirement that overcommit
-        // is not disabled (i.e., /proc/sys/vm/overcommit_memory != 2).
-        //
-        // Alternatively, we could use the more tedious but reliable approach of
-        // using mmap(PROT_NONE) and calling mprotect(PROT_READ|PROT_WRITE) on any
-        // pages we need to use (this is analogous to VirtualAlloc(MEM_RESERVE)
-        // followed by VirtualAlloc(MEM_COMMIT) in Windows).
-        common::map_fd_data(
-            this->m_data,
-            this->m_mapped_data_size,
-            PROT_READ | PROT_WRITE,
-            MAP_SHARED | MAP_NORESERVE,
-            this->m_fd,
-            0);
-
-        this->m_is_set = true;
-    }
-
-    // Opens a memory-mapped structure using a file descriptor.
-    //
-    // manage_fd is used to indicate whether the fd should be managed
-    // (i.e. closed at destruction time) by this class or not.
-    //
-    // Note: manage_fd also impacts the type of mapping: SHARED if true; PRIVATE otherwise.
-    // This is done for coding convenience because it suits current implementation,
-    // but could be changed in the future if we wish more control over this behavior.
-    void open(int fd, bool manage_fd = true)
-    {
-        common::retail_assert(
-            !this->m_is_set,
-            "Calling open() on an already set mapped_data_t instance!");
-
-        common::retail_assert(fd != -1, "mapped_data_t::open() was called with an invalid fd!");
-
-        this->m_mapped_data_size = sizeof(T);
-
-        if (manage_fd)
-        {
-            this->m_fd = fd;
-
-            common::map_fd_data(
-                this->m_data,
-                this->m_mapped_data_size,
-                PROT_READ | PROT_WRITE,
-                MAP_SHARED | MAP_NORESERVE,
-                this->m_fd,
-                0);
-
-            common::close_fd(this->m_fd);
-        }
-        else
-        {
-            common::map_fd_data(
-                this->m_data,
-                this->m_mapped_data_size,
-                PROT_READ | PROT_WRITE,
-                MAP_PRIVATE | MAP_NORESERVE,
-                fd,
-                0);
-        }
-
-        this->m_is_set = true;
-    }
-};
-
-// This class is similar to mapped_data_t, but is specialized for operation on log data structures.
-// There are enough differences from mapped_data_t to warrant a separate implementation.
-class mapped_log_t : public base_mapped_data_t<txn_log_t>
-{
-public:
-    mapped_log_t() = default;
-
-    // Copy semantics is disabled and moves should be performed via reset().
-    mapped_log_t(const mapped_log_t& other) = delete;
-    mapped_log_t(mapped_log_t&& other) = delete;
-    mapped_log_t& operator=(const mapped_log_t& rhs) = delete;
-    mapped_log_t& operator=(mapped_log_t&& rhs) = delete;
-
-    ~mapped_log_t() = default;
-
-    // Creates a memory-mapping for a log data structure.
-    void create(const char* name)
-    {
-        common::retail_assert(
-            !this->m_is_set,
-            "Calling create() on an already set mapped_log_t instance!");
-
-        this->m_fd = ::memfd_create(name, MFD_ALLOW_SEALING);
-        if (this->m_fd == -1)
-        {
-            common::throw_system_error("memfd_create() failed in mapped_log_t::create()!");
-        }
-
-        this->m_mapped_data_size = c_initial_log_size;
-
-        common::truncate_fd(this->m_fd, this->m_mapped_data_size);
-
-        common::map_fd_data(
-            this->m_data,
-            this->m_mapped_data_size,
-            PROT_READ | PROT_WRITE,
-            MAP_SHARED,
-            this->m_fd,
-            0);
-
-        this->m_is_set = true;
-    }
-
-    // Opens a memory-mapped log structure using a file descriptor.
-    void open(int fd)
-    {
-        common::retail_assert(
-            !this->m_is_set,
-            "Calling open() on an already set mapped_log_t instance!");
-
-        common::retail_assert(fd != -1, "mapped_log_t::open() was called with an invalid fd!");
-
-        this->m_mapped_data_size = common::get_fd_size(fd);
-
-        common::map_fd_data(
-            this->m_data,
-            this->m_mapped_data_size,
-            PROT_READ,
-            MAP_PRIVATE,
-            fd,
-            0);
-
-        this->m_is_set = true;
-    }
-
-    // Truncates and seals a memory-mapped log structure.
-    // Closes the mapped_log_t instance in the sense that it is left in an uninitialized state.
-    // The file descriptor is *NOT* closed - its ownership is transferred to the caller.
-    // Passes back the file descriptor and the size of the log.
-    void truncate_seal_and_close(int& fd, size_t& log_size)
-    {
-        common::retail_assert(
-            this->m_is_set,
-            "Calling truncate_seal_and_close() on an unset mapped_log_t instance!");
-
-        common::retail_assert(
-            this->m_fd != -1,
-            "truncate_seal_and_close() was called on a mapped_log_t instance that lacks a valid fd!");
-
-        fd = this->m_fd;
-        log_size = this->m_data->size();
-
-        common::unmap_fd_data(this->m_data, this->m_mapped_data_size);
-        this->m_mapped_data_size = 0;
-
-        common::truncate_fd(this->m_fd, log_size);
-
-        // Seal the txn log memfd for writes/resizing before sending it to the server.
-        if (-1 == ::fcntl(this->m_fd, F_ADD_SEALS, F_SEAL_SHRINK | F_SEAL_GROW | F_SEAL_WRITE))
-        {
-            common::throw_system_error(
-                "fcntl(F_ADD_SEALS) failed in mapped_log_t::truncate_seal_and_close()!");
-        }
-
-        this->m_fd = -1;
-
-        this->m_is_set = false;
-    }
-};
-
 } // namespace db
 } // namespace gaia
