+++ production/db/core/inc/db_server.hpp
@@ -19,7 +19,7 @@
 
 #include "gaia_internal/common/mmap_helpers.hpp"
 
-#include "db_internal_types.hpp"
+#include "mapped_data.hpp"
 #include "memory_manager.hpp"
 #include "messages_generated.h"
 #include "persistent_store_manager.hpp"
@@ -388,7 +388,8 @@ private:
     static std::pair<int, int> get_stream_socket_pair();
 
     template <typename element_type>
-    static void stream_producer_handler(int stream_socket, int cancel_eventfd, std::function<std::optional<element_type>()> generator_fn);
+    static void stream_producer_handler(
+        int stream_socket, int cancel_eventfd, std::function<std::optional<element_type>()> generator_fn);
 
     template <typename element_type>
     static void start_stream_producer(int stream_socket, std::function<std::optional<element_type>()> generator_fn);
@@ -536,65 +537,7 @@ private:
     class safe_fd_from_ts
     {
     public:
-        explicit safe_fd_from_ts(gaia_txn_id_t commit_ts, bool auto_close_fd = true)
-            : m_auto_close_fd(auto_close_fd)
-        {
-            common::retail_assert(is_commit_ts(commit_ts), "You must initialize safe_fd_from_ts from a valid commit_ts!");
-            // If the log fd was invalidated, it is either closed or soon will
-            // be closed, and therefore we cannot use it. We return early if the
-            // fd has already been invalidated to avoid the dup(2) call.
-            int log_fd = get_txn_log_fd(commit_ts);
-            if (log_fd == -1)
-            {
-                throw invalid_log_fd(commit_ts);
-            }
-
-            // To avoid races from the log fd being closed out from under us by a
-            // concurrent updater, we first dup(2) the fd, and then if the dup was
-            // successful, test the log fd entry again to ensure we aren't reusing a
-            // closed log fd. If the log fd was invalidated, then we need to close our
-            // dup fd and return false.
-            try
-            {
-                m_local_log_fd = common::duplicate_fd(log_fd);
-            }
-            catch (const common::system_error& e)
-            {
-                // The log fd was already closed by another thread (after being
-                // invalidated).
-                // NB: This does not handle the case of the log fd being closed and then
-                // reused before our dup(2) call. That case is handled below, where we
-                // check for invalidation.
-                if (e.get_errno() == EBADF)
-                {
-                    // The log fd must have been invalidated before it was closed.
-                    common::retail_assert(get_txn_log_fd(commit_ts) == -1, "log fd was closed without being invalidated!");
-                    // We lost the race because the log fd was invalidated and closed after our check.
-                    throw invalid_log_fd(commit_ts);
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            // If we were able to duplicate the log fd, check to be sure it
-            // wasn't invalidated (and thus possibly closed and reused before
-            // the call to dup(2)), so we know we aren't reusing a closed fd.
-            if (get_txn_log_fd(commit_ts) == -1)
-            {
-                // If we got here, we must have a valid dup fd.
-                common::retail_assert(
-                    common::is_fd_valid(m_local_log_fd),
-                    "fd should be valid if dup() succeeded!");
-                // We need to close the duplicated fd since the original fd
-                // might have been reused and we would leak it otherwise
-                // (because the destructor isn't called if the constructor
-                // throws).
-                common::close_fd(m_local_log_fd);
-                throw invalid_log_fd(commit_ts);
-            }
-        }
+        explicit safe_fd_from_ts(gaia_txn_id_t commit_ts, bool auto_close_fd = true);
 
         // The "rule of 3" doesn't apply here since we never pass this object
         // directly to a function; we always extract the safe fd value first.
@@ -602,26 +545,12 @@ private:
         safe_fd_from_ts(const safe_fd_from_ts&) = delete;
         safe_fd_from_ts& operator=(const safe_fd_from_ts&) = delete;
 
-        ~safe_fd_from_ts()
-        {
-            // Ensure we close the dup log fd. If the original log fd was closed
-            // already (indicated by get_txn_log_fd() returning -1), this will free
-            // the shared-memory object referenced by the fd.
-            if (m_auto_close_fd)
-            {
-                // If the constructor fails, this will handle an invalid fd (-1)
-                // correctly.
-                common::close_fd(m_local_log_fd);
-            }
-        }
+        ~safe_fd_from_ts();
 
         // This is the only public API. Callers are expected to call this method on
         // a stack-constructed instance of the class before they pass an fd to a
         // function that does not expect it to be concurrently closed.
-        int get_fd() const
-        {
-            return m_local_log_fd;
-        }
+        int get_fd() const;
 
     private:
         int m_local_log_fd{-1};
